requirements.txt
appnope==0.1.0
decorator==4.0.6
dj-database-url==0.3.0
Django==1.9
django-configurations==1.0
django-countries==3.4.1
django-debug-toolbar==1.4
django-email-log==0.2.0
django-extensions==1.6.1
django-jinja==2.1.1
django-localflavor==1.2
envdir==0.7
gunicorn==19.4.1
ipython==4.0.1
ipython-genutils==0.1.0
Jinja2==2.8
markdown2==2.3.0
MarkupSafe==0.23
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
PyYAML==3.11
setproctitle==1.1.9
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.1.18
traitlets==4.0.0
Werkzeug==0.11.3
wheel==0.24.0
